+++ test/integration/043_custom_aliases_test/macros-configs/macros.sql
@@ -9,7 +9,7 @@
 {%- endmacro %}
 
 {% macro string_literal(s) -%}
-  {{ adapter_macro('test.string_literal', s) }}
+  {{ adapter.dispatch('string_literal', packages=['test'])(s) }}
 {%- endmacro %}
 
 {% macro default__string_literal(s) %}